Move csv reader to inputlib
