+++ oemof/solph/__init__.py
@@ -7,4 +7,4 @@ from oemof.solph.network import (Sink, Source, LinearTransformer, Storage, Bus,
 from oemof.solph.models import OperationalModel
 from oemof.solph.groupings import GROUPINGS
 from oemof.solph.options import (Investment, BinaryFlow, DiscreteFlow)
-from oemof.solph.plumbing import NodesFromCSV
+from oemof.solph.inputlib.csv_tools import NodesFromCSV
